requirements.txt
pycryptodome
pyperclip
cryptography
tkinter
